// frontend/src/lib/api/models/SabbathClassResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SabbathClassResponseDto = {
    id: string;
    name: string;
    description?: string | null;
    displayOrder: number;
    isActive: boolean;
    createdAt: string;
    updatedAt?: string | null;
};